Remove redundant code in prepare
