+++ internal/runners/prepare/prepare.go
@@ -4,16 +4,20 @@ import (
 	"fmt"
 	"runtime"
 
+	svcAutostart "github.com/ActiveState/cli/cmd/state-svc/autostart"
+	trayAutostart "github.com/ActiveState/cli/cmd/state-tray/autostart"
 	"github.com/ActiveState/cli/internal/analytics"
 	"github.com/ActiveState/cli/internal/captain"
 	"github.com/ActiveState/cli/internal/config"
 	"github.com/ActiveState/cli/internal/constants"
 	"github.com/ActiveState/cli/internal/errs"
 	"github.com/ActiveState/cli/internal/globaldefault"
+	"github.com/ActiveState/cli/internal/installation"
 	"github.com/ActiveState/cli/internal/installation/storage"
 	"github.com/ActiveState/cli/internal/locale"
 	"github.com/ActiveState/cli/internal/logging"
 	"github.com/ActiveState/cli/internal/multilog"
+	"github.com/ActiveState/cli/internal/osutils/autostart"
 	"github.com/ActiveState/cli/internal/output"
 	"github.com/ActiveState/cli/internal/primer"
 	"github.com/ActiveState/cli/internal/subshell"
@@ -144,4 +148,51 @@ func updateConfigKey(cfg *config.Instance, oldKey, newKey string) error {
 	}
 
 	return nil
+}
+
+// InstalledPreparedFiles returns the files installed by state _prepare
+func InstalledPreparedFiles(cfg autostart.Configurable) ([]string, error) {
+	var files []string
+	trayExec, err := installation.TrayExec()
+	if err != nil {
+		return nil, locale.WrapError(err, "err_tray_exec")
+	}
+
+	trayShortcut, err := autostart.New(trayAutostart.App, trayExec, nil, trayAutostart.Options, cfg)
+	if err != nil {
+		return nil, locale.WrapError(err, "err_autostart_app")
+	}
+
+	path, err := trayShortcut.Path()
+	if err != nil {
+		multilog.Error("Failed to determine shortcut path for removal: %v", err)
+	} else if path != "" {
+		files = append(files, path)
+	}
+
+	svcExec, err := installation.ServiceExec()
+	if err != nil {
+		return nil, locale.WrapError(err, "err_svc_exec")
+	}
+
+	svcShortuct, err := autostart.New(svcAutostart.App, svcExec, []string{"start"}, svcAutostart.Options, cfg)
+	if err != nil {
+		return nil, locale.WrapError(err, "err_autostart_app")
+	}
+
+	path, err = svcShortuct.Path()
+	if err != nil {
+		multilog.Error("Failed to determine shortcut path for removal: %v", err)
+	} else if path != "" {
+		files = append(files, path)
+	}
+
+	osSpecificFiles, err := installedPreparedFiles(cfg)
+	if err != nil {
+		return nil, locale.WrapError(err, "err_prepare_os_files", "Could not get list of OS specific prepared files")
+	}
+
+	files = append(files, osSpecificFiles...)
+
+	return files, nil
 }
