+++ internal/runners/prepare/prepare_darwin.go
@@ -35,42 +35,6 @@ func (r *Prepare) prepareOS() error {
 	return nil
 }
 
-// InstalledPreparedFiles returns the files installed by the prepare command
-func InstalledPreparedFiles(cfg autostart.Configurable) ([]string, error) {
-	var files []string
-	trayExec, err := installation.TrayExec()
-	if err != nil {
-		return nil, locale.WrapError(err, "err_tray_exec")
-	}
-
-	sc, err := autostart.New(trayAutostart.App, trayExec, nil, trayAutostart.Options, cfg)
-	if err != nil {
-		return nil, locale.WrapError(err, "err_autostart_app")
-	}
-
-	path, err := sc.Path()
-	if err != nil {
-		multilog.Error("Failed to determine shortcut path for removal: %v", err)
-	} else if path != "" {
-		files = append(files, path)
-	}
-
-	svcExec, err := installation.ServiceExec()
-	if err != nil {
-		return nil, locale.WrapError(err, "err_svc_exec")
-	}
-
-	sc, err = autostart.New(svcAutostart.App, svcExec, []string{"start"}, svcAutostart.Options, cfg)
-	if err != nil {
-		return nil, locale.WrapError(err, "err_autostart_app")
-	}
-
-	path, err = sc.Path()
-	if err != nil {
-		multilog.Error("Failed to determine shortcut path for removal: %v", err)
-	} else if path != "" {
-		files = append(files, path)
-	}
-
-	return files, nil
+func installedPreparedFiles(cfg autostart.Configurable) ([]string, error) {
+	return nil, nil
 }
